FWManagement/requirements.txt
caldav==0.7.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
decorator==4.4.2
dnspython==2.0.0
dominate==2.5.2
email-validator==1.1.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-WTF==0.14.3
icalendar==4.0.7
idna==2.10
infinity==1.5
intervals==0.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
ldap3==2.8.1
lxml==4.5.2
MarkupSafe==1.1.1
phonenumbers==8.12.9
pyasn1==0.4.8
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
validators==0.18.1
visitor==0.1.3
vobject==0.9.6.1
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-Components==0.10.4
xlwt==1.3.0
